+++ src/Pim/Bundle/CatalogBundle/Doctrine/Common/Remover/CategoryRemover.php
@@ -13,7 +13,9 @@ use Pim\Bundle\CatalogBundle\Model\CategoryInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 /**
- * Category remover
+ * TODO: To extract in a component
+ *
+ * Product category remover
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)