+++ src/Pim/Bundle/CatalogBundle/PimCatalogBundle.php
@@ -8,6 +8,7 @@ use Oro\Bundle\EntityBundle\DependencyInjection\Compiler\DoctrineOrmMappingsPass
 use Pim\Bundle\CatalogBundle\DependencyInjection\Compiler\RegisterAttributeConstraintGuessersPass;
 use Pim\Bundle\CatalogBundle\DependencyInjection\Compiler\RegisterAttributeTypePass;
 use Pim\Bundle\CatalogBundle\DependencyInjection\Compiler\RegisterComparatorsPass;
+use Pim\Bundle\CatalogBundle\DependencyInjection\Compiler\RegisterCompleteCheckerPass;
 use Pim\Bundle\CatalogBundle\DependencyInjection\Compiler\RegisterFilterPass;
 use Pim\Bundle\CatalogBundle\DependencyInjection\Compiler\RegisterProductQueryFilterPass;
 use Pim\Bundle\CatalogBundle\DependencyInjection\Compiler\RegisterProductQuerySorterPass;
@@ -41,7 +42,8 @@ class PimCatalogBundle extends Bundle
             ->addCompilerPass(new RegisterProductQuerySorterPass())
             ->addCompilerPass(new RegisterProductUpdaterPass())
             ->addCompilerPass(new RegisterFilterPass())
-            ->addCompilerPass(new RegisterComparatorsPass());
+            ->addCompilerPass(new RegisterComparatorsPass())
+            ->addCompilerPass(new RegisterCompleteCheckerPass());
 
         $productMappings = [
             realpath(__DIR__ . '/Resources/config/model/doctrine') => 'Pim\Bundle\CatalogBundle\Model'