+++ src/Pim/Bundle/DataGridBundle/Extension/Filter/FilterExtension.php
@@ -198,18 +198,35 @@ class FilterExtension extends AbstractExtension
             $filters[] = $this->getFilterObject($column, $filter);
         }
 
+        // TODO: Try to make filter without views, to remove this kind of stuff
         $gridName = $config->offsetGetByPath('[name]');
-        $isProductGrid = preg_match('/product-grid$/', $gridName);
+        $gridCategoryConfig = $this->getCategoryFilterConfig($gridName);
 
-        if (!isset($filtersConfig['category']) && $isProductGrid) {
-            $categoryConfig = [
-                'type'      => 'product_category',
-                'data_name' => 'category',
-            ];
-            $filters[] = $this->getFilterObject('category', $categoryConfig);
+        if (!isset($filtersConfig['category']) && null !== $gridCategoryConfig) {
+            $filters[] = $this->getFilterObject('category', $gridCategoryConfig);
         }
 
         return $filters;
+    }
+
+    /**
+     * Return the category filter config for the given $gridname,
+     * if this $gridname is not filterable by category, return null.
+     *
+     * @param $gridname
+     *
+     * @return array|null
+     */
+    protected function getCategoryFilterConfig($gridname)
+    {
+        $gridConfigs = [
+            'product-grid' => [
+                'type'      => 'product_category',
+                'data_name' => 'category'
+            ]
+        ];
+
+        return isset($gridConfigs[$gridname]) ? $gridConfigs[$gridname] : null;
     }
 
     /**