+++ src/Pim/Bundle/UIBundle/Tests/Unit/Controller/AjaxOptionControllerTest.php
@@ -42,7 +42,8 @@ class AjaxOptionControllerTest extends \PHPUnit_Framework_TestCase
             'dataLocale'    => 'data_locale',
             'collectionId'  => 'collection_id',
             'search'        => 'search',
-            'options'       => array('options')
+            'options'       => array('options'),
+            'isCreatable'   => false
         );
         $request->query->expects($this->any())
             ->method('get')